pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "serial_weighing_scale"
version = "2.0.0"
authors = [
    { name = "Lars B. Rollik", email = "L.B.Rollik@protonmail.com" }
]
license = { file = "LICENSE" }
description = "serial_weighing_scale: Python API for serial weighing scales based on Arduino+HX711"
readme = { file = "README.md", content-type = "text/markdown" }
classifiers = [
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "License :: OSI Approved :: BSD License",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.6",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
requires-python = ">=3.6"
dependencies = [
    "pyserial",
    # struct, json are built-in
]

[project.urls]
Repository = "https://github.com/larsrollik/serial_weighing_scale"
"Issue Tracker" = "https://github.com/larsrollik/serial_weighing_scale/issues"


[project.optional-dependencies]
dev = [
    "black",
    "bump2version",
    "pytest-cov",  # implies: pytest, coverage dependencies
    "pre-commit",
    "toml",
    "types-PyYAML",
]

[project.entry-points.console_scripts]
#serial_weighing_scale-entrypoint = "serial_weighing_scale.__init__:run"

[tool.setuptools]
zip-safe = false
include-package-data = true

[tool.setuptools.packages.find]
where = [""]

[tool.setuptools.package-data]
"*" = ["*.config"]

[tool.pytest.ini_options]
addopts = "--cov=serial_weighing_scale --cov-report=term-missing --maxfail=5"

[tool.ruff]
line-length = 99
indent-width = 4
include = ["*.pyi"]
extend-exclude = ["*.eggs", "*.git", "*.hg", "*.mypy_cache", "*.tox", "*.venv", "*_build", "*buck-out", "*build", "*dist"]

[tool.ruff.lint]
select = [
    "E",   # Pyflakes (general errors)
    "W",   # Pycodestyle warnings
    "F",   # Pyflakes checks (includes unused imports)
    "I",   # isort (import sorting)
    "UP",  # pyupgrade (Python syntax upgrades)
    "YTT", # check for YAML issues
    "SIM", # detect unnecessary comprehensions, loops, etc.
    "PTH", # checks for pathlib best practices
    "TCH", # checks for typing best practices
    "PYI", # checks for pyi files
]
ignore = []

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"
# Like Black, indent with spaces, rather than tabs.
indent-style = "space"
# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false
# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

# Enable auto-formatting of code examples in docstrings. Markdown,
# reStructuredText code/literal blocks and doctests are all supported.
#
# This is currently disabled by default, but it is planned for this
# to be opt-out in the future.
docstring-code-format = true
docstring-code-line-length = "dynamic"

[tool.ruff.lint.pydocstyle]
convention = "numpy"  # Accepts: "google", "numpy", or "pep257".

[tool.ruff.lint.per-file-ignores]
".bumpversion.cfg" = ["E501", "W291", "W292"]  # Equivalent to excluding it from whitespace checks

[tool.mypy]
# strict = true
# follow_imports = "silent"
warn_return_any = false
ignore_missing_imports = true
